pyproject.toml
[tool.poetry]
name = "flatpak-rebuilder"
version = "0.1.0"
description = "A script to automatically rebuild a flatpak."
authors = ["Zacharie Tevaearai <zacharie.tevaearai@epfl.ch>"]

[tool.poetry.scripts]
flatpak-rebuilder = 'flatpak_rebuilder.main:main'

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
